third_party/blink/web_tests/fast/speechsynthesis/speech-synthesis-pause-resume-expected.txt

This tests that pausing/resuming speech jobs works as expected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

On pause event received.
PASS speechSynthesis.paused is true
PASS speechSynthesis.speaking is true
On resume event received.
PASS speechSynthesis.paused is false
PASS speechSynthesis.speaking is true
On end event received.
PASS speechSynthesis.paused is false
PASS speechSynthesis.speaking is false
PASS successfullyParsed is true

TEST COMPLETE
